projects/experimental/NS_0327/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hr-admin-bots"
version = "0.5.0"
description = "HR Admin Bot Skill Package — Telegram bots for onboarding, work permit, leave, and offboarding"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot>=20.0",
    "gspread>=5.0",
    "google-auth>=2.0",
]

[project.scripts]
hr-admin-bots = "hr_admin_bots.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
